@@ -12,9 +12,12 @@
     "@fortawesome/fontawesome-svg-core": "^6.4.2",
     "@fortawesome/free-solid-svg-icons": "^6.4.2",
     "@fortawesome/react-fontawesome": "^0.2.0",
-    "next": "13.5.6",
+    "bootstrap": "^5.3.2",
+    "next": "^14.0.1",
     "react": "^18",
-    "react-dom": "^18"
+    "react-bootstrap": "^2.9.1",
+    "react-dom": "^18",
+    "react-responsive-carousel": "^3.2.23"
   },
   "devDependencies": {
     "eslint": "^8",
