@@ -1,3 +1,6 @@
+import "bootstrap/dist/css/bootstrap.css";
+import "react-responsive-carousel/lib/styles/carousel.min.css";
+import BootStrapJs from "./components/BootstrapJs";
 import { Inter } from "next/font/google";
 import "./globals.css";
 
@@ -11,7 +14,10 @@ export const metadata = {
 export default function RootLayout({ children }) {
   return (
     <html lang="en">
-      <body className={inter.className}>{children}</body>
+      <body className={inter.className}>
+        {children}
+        <BootStrapJs />
+      </body>
     </html>
   );
 }
